* Refactor Model-like classes out into Models folder * Refactor classes out into separate class files * Removal of redundant generics * Change OrderBy-OrderBy to OrderBy-ThenBy * Removal of unused using references * Updates to namespaces based on refactoring * Conversion of String.Format to C# 6.0 String-interpolation
** Ensured all existing Core Unit Tests still green.
