+++ EveryDayShuffling/Models/CardValue.cs
@@ -1,0 +1,19 @@
+﻿namespace EveryDayShuffling.Models
+{
+    public enum CardValue
+    {
+        Two = 2,
+        Three = 3,
+        Four = 4,
+        Five = 5,
+        Six = 6,
+        Seven = 7,
+        Eight = 8,
+        Nine = 9,
+        Ten = 10,
+        Jack = 11,
+        Queen = 12,
+        King = 13,
+        Ace = 1
+    }
+}
