+++ EveryDayShuffling/Models/Suit.cs
@@ -1,0 +1,11 @@
+﻿
+namespace EveryDayShuffling.Models
+{
+    public enum Suit
+    {
+        Club = 1,
+        Diamond = 2,
+        Heart = 3,
+        Spade = 4
+    }
+}
